move common methods to renderHTML
