+++ src/renderHTML.js
@@ -26,9 +26,6 @@ md.use(mdemoji);
 // set mdemoji rules
 md.renderer.rules.emoji = (token, idx) => `<i class="twa twa-${token[idx].markup}"></i>`;
 
-const defaultTemplatePath = 'src/template/web/index.html';
-
-
 /**
  * Using the given parameters, calls the Mustache engine
  * and renders the HTML page.
@@ -57,4 +54,60 @@ exports.transformTemplate = (
     // calls the render engine
     const output = Mustache.render(templateContent, view);
     return output;
+};
+
+exports.renderLicense = (
+    templateContent, contractsStructure,
+) => {
+    const LICENSEText = String(fs.readFileSync(path.join(process.cwd(), 'LICENSE'))).trim();
+    const LICENSE = LICENSEText.replace(/\n/g, '<br>');
+    // put all data together
+    const view = {
+        contractsStructure,
+        hasLICENSE: true,
+        LICENSE,
+    };
+    // calls the render engine
+    return Mustache.render(templateContent, view);
+};
+
+exports.renderReadme = (
+    templateContent, contractsStructure, hasLICENSE,
+) => {
+    const READMEText = String(fs.readFileSync(path.join(process.cwd(), 'README.md'))).trim();
+    // render it, from markdown to html
+    const READMEconverted = md.render(READMEText);
+    const README = READMEconverted
+        .replace(/<h1>/g, '<h1 class="title is-1">')
+        .replace(/<h2>/g, '<h2 class="title is-2">')
+        .replace(/<h3>/g, '<h3 class="title is-3">')
+        .replace(/<h4>/g, '<h4 class="title is-4">')
+        .replace(/<ul>/g, '<ul class="menu-list">');
+    // put all data together
+    const view = {
+        contractsStructure,
+        hasLICENSE,
+        README,
+    };
+    // calls the render engine
+    return Mustache.render(templateContent, view);
+};
+
+exports.organizeContractsStructure = (
+    contractsPreparedData,
+) => {
+    const contractsStructure = [];
+    contractsPreparedData.forEach((contract) => {
+        const contractInfo = {};
+        // add name
+        contractInfo.name = contract.contractName;
+        contractInfo.filename = contract.filename;
+        contractInfo.functions = [];
+        // add functions name
+        contract.contractData.functions.forEach((func) => {
+            contractInfo.functions.push({ name: func.ast.name });
+        });
+        contractsStructure.push(contractInfo);
+    });
+    return contractsStructure;
 };
